refactor: 기존 쿠폰 리스트 섹션과 유사하게 폴더링 수정
